pyproject.toml
[tool.poetry]
name = "ai-assistant"
version = "0.1.0"
description = ""
authors = ["Priyadarshi Roy <proy9714@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyrect = "^0.2.0"
python-dotenv = "^1.0.1"
pynput = "^1.7.6"

[tool.poetry.group.dev.dependencies]
pyperclip = "^1.8.2"
langchain = "^0.1.9"
langchain-google-genai = "^0.0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
